feat: refactor logging implementation and enhance error handling in transfer scripts
